+++ packages/skill-search/index.js
@@ -40,5 +40,7 @@ export {
   skillSearchNpm,
   skillSearchStackoverflow,
   skillSearchCheatsh,
+  searchSkills,
+  devSkills,
   skills,
 };